deleted schedule service
